+++ index.html
@@ -38,13 +38,11 @@
                 fill="#879298"
               />
             </svg>
-            <!-- <span>리스트형</span> -->
           </button>
           <button class="grid-btn">
             <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none">
               <path d="M3 11V3H11V11H3ZM3 21V13H11V21H3ZM13 11V3H21V11H13ZM13 21V13H21V21H13ZM5 9H9V5H5V9ZM15 9H19V5H15V9ZM15 19H19V15H15V19ZM5 19H9V15H5V19Z" fill="#4362D0" />
             </svg>
-            <!-- <span>그리드형</span> -->
           </button>
         </div>
       </nav>
